# api/openapi.yaml
openapi: 3.1.0
info:
  title: Wuling DevOps API
  version: 0.1.0-stage1
  description: |
    Stage 1 of the Wuling DevOps platform.

    What's implemented:
      - Auth (register, login, /me, personal access tokens, SSH public keys)
      - Orgs and Projects (CRUD)
      - Repositories (CRUD + libgit2-backed read APIs)
      - Git smart HTTP (clone/fetch/push) and Git over SSH (default :2222)
      - Issues (CRUD, labels, comments, assignees)
      - Merge Requests (open/diff/merge with FF/merge-commit/squash, comments, reviews)
      - Wiki (per-project Markdown pages backed by a separate Git repo)
      - Insights (per-project activity, contributors, language stats)

    Planned for later stages (URLs reserved, not yet served):
      - Pipelines

    All errors use the JSON envelope:
      { "error": { "code": "<stable_code>", "message": "<human msg>", "details": {} } }
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - bearerAuth: []
paths:
...
  /api/v1/orgs/{org_slug}/members:
    get:
      summary: List members of an org
      parameters:
        - in: path
          name: org_slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items: { $ref: "#/components/schemas/OrgMember" }
                  role:
                    $ref: "#/components/schemas/OrgRole"
        "404": { $ref: "#/components/responses/NotFoundError" }
  /api/v1/orgs/{org_slug}/members/{user_id}:
    patch:
      summary: Change a member's role
      description: |
        Owners may grant any role including owner. Maintainers may grant
        developer/reporter/guest but never maintainer/owner. Demoting the
        last owner is refused (409).
      parameters:
        - in: path
          name: org_slug
          required: true
          schema: { type: string }
        - in: path
          name: user_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchMemberRequest" }
      responses:
        "204": { description: updated }
        "403": { $ref: "#/components/responses/ForbiddenError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "409":
          description: Conflict — for example, refusing to demote the last owner.
    delete:
      summary: Remove a member from an org (or self-leave)
      description: |
        Self-leave (user_id == caller) is permitted regardless of role,
        subject to the last-owner guard. Removing another member requires
        manage-members permission and a strictly higher rank than the target.
      parameters:
        - in: path
          name: org_slug
          required: true
          schema: { type: string }
        - in: path
          name: user_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: removed }
        "403": { $ref: "#/components/responses/ForbiddenError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "409":
          description: Conflict — last owner cannot leave.

  # ---------------- invitations ----------------
...
  /api/v1/orgs:
    get:
      summary: List orgs the caller belongs to
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  orgs:
                    type: array
                    items: { $ref: "#/components/schemas/Org" }
    post:
      summary: Create a new (non-personal) org; caller becomes its owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug]
              properties:
                slug: { type: string, minLength: 2, maxLength: 64 }
                display_name: { type: string, maxLength: 128 }
                description: { type: string, maxLength: 512 }
      responses:
        "201":
          description: created
          content: { application/json: { schema: { $ref: "#/components/schemas/Org" } } }
        "409": { $ref: "#/components/responses/ConflictError" }
  /api/v1/orgs/{org_slug}:
    parameters:
      - $ref: "#/components/parameters/OrgSlug"
    get:
      summary: Fetch an org by slug
      responses:
        "200":
          description: ok
          content: { application/json: { schema: { $ref: "#/components/schemas/Org" } } }
        "404": { $ref: "#/components/responses/NotFoundError" }
...
  /api/v1/orgs/{org_slug}/projects/{project_slug}/wiki/pages:
    parameters:
      - $ref: "#/components/parameters/OrgSlug"
      - $ref: "#/components/parameters/ProjectSlug"
    get:
      summary: List every Markdown page in the project wiki
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  pages:
                    type: array
                    items: { $ref: "#/components/schemas/WikiPage" }
        "404": { $ref: "#/components/responses/NotFoundError" }
...
  /api/v1/orgs/{org_slug}/projects/{project_slug}/wiki/history:
    parameters:
      - $ref: "#/components/parameters/OrgSlug"
      - $ref: "#/components/parameters/ProjectSlug"
      - in: query
        name: limit
        schema: { type: integer, minimum: 1, maximum: 200 }
    get:
      summary: Recent commits on the wiki repo's default branch
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  commits:
                    type: array
                    items: { $ref: "#/components/schemas/GitCommit" }

  # ---------------- insights ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
      description: |
        Used by Git CLI only. The "password" is either a personal access token
        (preferred — recognised by the `wlpat_` prefix) or the user's account
        password.

        **Security:** credentials are sent on every request, so basic auth
        MUST be used over HTTPS in production — configure server URLs as
        `https://` and terminate TLS in front of the API. For local
        development against `localhost` plain HTTP is acceptable.

        Operators should prefer personal access tokens (`wlpat_…`) over
        account passwords, and consider OAuth/bearer-token flows for
        non-CLI clients.
  parameters:
    OrgSlug:
      in: path
      name: org_slug
      required: true
      schema: { type: string }
    ProjectSlug:
      in: path
      name: project_slug
      required: true
      schema: { type: string }
...
  responses:
...
    ForbiddenError:
      description: caller lacks permission
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFoundError:
      description: resource not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    ConflictError:
      description: resource already exists
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - validation
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - already_exists
                - rate_limited
                - unsupported
                - bad_request
                - internal
                - unavailable
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    Org:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        display_name: { type: string }
        description: { type: string }
        is_personal: { type: boolean }
        created_at: { type: string, format: date-time }
    OrgRole:
      type: string
      enum: [owner, maintainer, developer, reporter, guest]
      description: |
        Org-level membership role, GitLab-style. owner > maintainer >
        developer > reporter > guest. Permission helpers in internal/auth
        (CanReadOrg / CanWriteRepo / CanModerateContent / CanManageMembers /
        CanAdminOrg) gate every action against this enum.
...
    OrgMember:
      type: object
      properties:
        user_id:      { type: string, format: uuid }
        username:     { type: string }
        display_name: { type: string }
        email:        { type: string }
        avatar_url:   { type: string }
        role:         { $ref: "#/components/schemas/OrgRole" }
        joined_at:    { type: string, format: date-time }
...
    PatchMemberRequest:
      type: object
      required: [role]
      properties:
        role: { $ref: "#/components/schemas/OrgRole" }
...
    WikiPage:
      type: object
      properties:
        path: { type: string }
        size: { type: integer, format: int64 }
        updated_at: { type: string, format: date-time, nullable: true }
...
    GitCommit:
      type: object
      description: A single commit as returned by the libgit2 wrapper.
      properties:
        oid: { type: string }
        tree_oid: { type: string }
        message: { type: string }
        parents:
          type: array
          items: { type: string }
        author:
          type: object
          properties:
            name: { type: string }
            email: { type: string }
            when: { type: string, format: date-time }
        committer:
          type: object
          properties:
            name: { type: string }
            email: { type: string }
            when: { type: string, format: date-time }

    # ---------------- Insights ----------------
...